+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Fallback to primitive string type for malformed JSON and JSON5 inputs
 - Add more details to exception messages
 
+### Fixed
+
+- PsiInvalidElementAccessException on content reload
+
 ## [5.0.1] - 2026-04-19
 
 ### Changed